sample_clients/python/requirements.txt
kafka-python
krbcontext
krbticket
gssapi
